+++ ubelt/util_cmd.py
@@ -24,8 +24,9 @@ Example:
 """
 from __future__ import absolute_import, division, print_function, unicode_literals
 import sys
-import six
 import warnings
+
+__all__ = ['cmd']
 
 POSIX = 'posix' in sys.builtin_module_names
 
@@ -34,21 +35,13 @@ if POSIX:
 else:  # nocover
     select = NotImplemented
 
-__all__ = ['cmd']
-
-# def _run_process(proc):
-#     """ helper for cmd """
-#     while True:
-#         # returns None while subprocess is running
-#         retcode = proc.poll()
-#         line = proc.stdout.readline()
-#         yield line
-#         if retcode is not None:
-#             # The program has a return code, so its done executing.
-#             # Grab any remaining data in stdout
-#             for line in proc.stdout.readlines():
-#                 yield line
-#             raise StopIteration('process finished')
+PY2 = sys.version_info[0] == 2
+PY34 = sys.version_info[0:2] == (3, 4)
+if PY2:
+    import six
+    string_types = six.string_types
+else:
+    string_types = (str,)
 
 
 def _textio_iterlines(stream):
@@ -331,7 +324,7 @@ def cmd(command, shell=False, detach=False, verbose=0, tee=None, cwd=None,
             raise ValueError('Unknown kwargs: {}'.format(list(kwargs.keys())))
 
     # Determine if command is specified as text or a tuple
-    if isinstance(command, six.string_types):
+    if isinstance(command, string_types):
         command_text = command
         command_tup = None
     else:
@@ -425,7 +418,7 @@ def cmd(command, shell=False, detach=False, verbose=0, tee=None, cwd=None,
 
         if check:
             if info['ret'] != 0:
-                if six.PY2 or six.PY34:
+                if PY2 or PY34:
                     raise subprocess.CalledProcessError(
                         info['ret'], info['command'], info['out'])
                 else: